# swagger.json
{"openapi":"3.0.0","info":{"title":"Coffee Addiction API","version":"1.0.0","description":"Coffee Addiction API swagger documentation"},"servers":[{"url":"http://localhost:3000"}],"paths":{"/beans":{"get":{"summary":"Returns list of coffee beans","description":"Retrieves a list of coffee beans from the database","tags":["Coffee Beans"],"responses":{"200":{"description":"Successful response","content":{"application/json":{"schema":{"type":"object","properties":{"beans":{"type":"array","items":{"type":"object","properties":{"beanId":{"type":"integer","description":"Unique identifier for the coffee bean"},"name":{"type":"string","description":"Name of the coffee bean"},"origin":{"type":"string","description":"Country or region of origin"},"roastLevel":{"type":"string","description":"Level of roast (e.g., light, medium, dark)"},"imageUrl":{"type":"string","description":"URL"},"pricePerKg":{"type":"number","format":"float","description":"Price per kilogram of coffee beans"},"stockQuantity":{"type":"integer","description":"Available quantity in stock"},"description":{"type":"string","description":"Detailed description of the coffee bean"}}}}}}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string","description":"Error message"}}}}}}}},"post":{"summary":"Creates a new bean","description":"Creates a new bean in the database","tags":["Coffee Beans"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string","description":"Name of the bean"},"origin":{"type":"string","description":"Country or region of origin"},"roastLevel":{"type":"string","description":"Level of roast (e.g., light, medium, dark)"},"pricePerKg":{"type":"number","format":"float","description":"Price per kilogram of coffee beans"},"imageUrl":{"type":"string","description":"URL"},"stockQuantity":{"type":"integer","description":"Available quantity in stock"},"description":{"type":"string","description":"Detailed description of the coffee bean"}}}}}},"responses":{"200":{"description":"Successful response","content":{"application/json":{"schema":{"type":"object","properties":{"bean":{"type":"object","properties":{"bean_id":{"type":"integer","description":"Unique identifier for the coffee bean"},"name":{"type":"string","description":"Name of the coffee bean"},"origin":{"type":"string","description":"Country or region of origin"},"roastLevel":{"type":"string","description":"Level of roast (e.g., light, medium, dark)"},"pricePerKg":{"type":"number","format":"float","description":"Price per kilogram of coffee beans"},"imageUrl":{"type":"string","description":"URL"},"stockQuantity":{"type":"integer","description":"Available quantity in stock"},"description":{"type":"string","description":"Detailed description of the coffee bean"}}}}}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string","description":"Error message"}}}}}}}}},"/login":{"post":{"summary":"Logs in a user","description":"Logs in a user with their email and password","tags":["Users"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"username":{"type":"string","description":"User's username"},"password":{"type":"string","description":"Password of the user"}}}}}},"responses":{"200":{"description":"Successful response","content":{"application/json":{"schema":{"type":"object","properties":{"userId":{"type":"integer","description":"Unique identifier for the user"},"username":{"type":"string","description":"Username of the user"},"jwt":{"type":"string","description":"JWT token of user"},"loginAttempts":{"type":"integer","description":"Number of login attempts for the user"},"firstName":{"type":"string","description":"First name of the user"},"lastName":{"type":"string","description":"Last name of the user"},"createdAt":{"type":"string","format":"date-time","description":"Date and time when the user was created"}}}}}},"400":{"description":"Invalid request","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string","description":"Error message"}}}}}}}}},"/signup":{"post":{"summary":"Signs up a new user","description":"Signs up a new user with their email, password, and other details","tags":["Users"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"username":{"type":"string","description":"Desired username"},"password":{"type":"string","description":"Password of the user"},"firstName":{"type":"string","description":"First name of the user"},"lastName":{"type":"string","description":"Last name of the user"}}}}}},"responses":{"200":{"description":"Successful response","content":{"application/json":{"schema":{"type":"object","properties":{"jwt":{"type":"string","description":"JSON Web Token"}}}}}},"400":{"description":"Invalid request","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string","description":"Error message"}}}}}}}}},"/user/beans":{"get":{"summary":"Returns list of beans owned by the user","description":"Retrieves a list of beans owned by the user from the database","tags":["Users"],"responses":{"200":{"description":"Successful response","content":{"application/json":{"schema":{"type":"object","properties":{"beans":{"type":"array","items":{"type":"object","properties":{"beanId":{"type":"integer","description":"Unique identifier for the coffee bean"},"name":{"type":"string","description":"Name of the coffee bean"},"origin":{"type":"string","description":"Country or region of origin"},"roastLevel":{"type":"string","description":"Level of roast (e.g., light, medium, dark)"},"imageUrl":{"type":"string","description":"URL"},"pricePerKg":{"type":"number","format":"float","description":"Price per kilogram of coffee beans"},"stockQuantity":{"type":"integer","description":"Available quantity in stock"},"description":{"type":"string","description":"Detailed description of the coffee bean"}}}}}}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string","description":"Error message"}}}}}}}}},"/user/bean/create":{"post":{"summary":"Creates a new bean owned by the user","description":"Creates a new bean owned by the user in the database","tags":["Users"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string","description":"Name of the bean"},"origin":{"type":"string","description":"Country or region of origin"},"roastLevel":{"type":"string","description":"Level of roast (e.g., light, medium, dark)"},"pricePerKg":{"type":"number","format":"float","description":"Price per kilogram of coffee beans"},"imageUrl":{"type":"string","description":"URL"},"stockQuantity":{"type":"integer","description":"Available quantity in stock"},"description":{"type":"string","description":"Detailed description of the coffee bean"}}}}}},"responses":{"200":{"description":"Successful response","content":{"application/json":{"schema":{"type":"object","properties":{"bean":{"type":"object","properties":{"bean_id":{"type":"integer","description":"Unique identifier for the coffee bean"},"name":{"type":"string","description":"Name of the coffee bean"},"origin":{"type":"string","description":"Country or region of origin"},"roastLevel":{"type":"string","description":"Level of roast (e.g., light, medium, dark)"},"pricePerKg":{"type":"number","format":"float","description":"Price per kilogram of coffee beans"},"imageUrl":{"type":"string","description":"URL"},"stockQuantity":{"type":"integer","description":"Available quantity in stock"},"description":{"type":"string","description":"Detailed description of the coffee bean"}}}}}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string","description":"Error message"}}}}}}}}},"/bean/{id}":{"delete":{"summary":"Deletes a bean owned by the user","description":"Deletes a bean owned by the user from the database","tags":["Users"],"parameters":[{"in":"path","name":"id","schema":{"type":"integer"},"required":true}],"responses":{"200":{"description":"Successful response","content":{"application/json":{"schema":{"type":"object","properties":{"result":{"type":"boolean","description":"Whether the bean was deleted successfully"}}}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string","description":"Error message"}}}}}}}}}},"components":{},"tags":[]}